Major refactoring of the markup and javascript rendering engine.
Widgets are now compiled to Markup at the end of the request, during the
request they are only scheduled for rendering. This means that when you
add a new Widget to the UI, the Widget is not immediately compiled. The
advantage is you can continue to style your Widget, add listeners... and
they will be compiled to Markup at the end of the request. This avoids a
bunch of javascript and jQuery calls when the component has already been
rendered.

A first unit test for the rendering engine has been included in
org.whizu.widget.ContainerTest.